Clean up PIP-CO Startup Systems modular update
