+++ holotapes/PipCoStartupSystems/app.js
@@ -1,0 +1,489 @@
+(function () {
+  const fs = require('fs'),
+    B = 'HOLO/STARTUP_ANIMATIONS/',
+    S = B + 'SELECT.JSON';
+  let c = 0,
+    s = 0,
+    t,
+    page = 0,
+    ic = [0, 0, 0, 0, 0, 0, 0, 0, 0];
+  let ownVideo = 0,
+    ownAudio = 0,
+    aHook,
+    bHook;
+  let TI = {
+    title: {
+      width: 176,
+      height: 98,
+      bpp: 1,
+      transparent: 0,
+      buffer: fs.readFileSync(B + 'TITLE.BIN'),
+    },
+  };
+  function customHookAlive() {
+    return (
+      global.__startupAnimationsHoloLiveV5 ||
+      global.__startupAnimationsHoloLiveV6 ||
+      global.__startupAnimationsHoloLiveV7 ||
+      global.__startupAnimationsHoloLiveV8 ||
+      global.__startupAnimationsHoloLiveV9 ||
+      global.__startupAnimationsHoloLiveV12 ||
+      global.__startupAnimationsHoloLiveV14 ||
+      global.__SA16 ||
+      global.__SA17 ||
+      global.__SA18 ||
+      global.__SA22
+    );
+  }
+  if (!customHookAlive()) {
+    if (!Pip.__SA22CleanBoot) Pip.__SA22CleanBoot = Pip.bootAnimation.bind(Pip);
+    if (!Pip.__SA22CleanAudio) Pip.__SA22CleanAudio = Pip.audioStart.bind(Pip);
+  }
+  function ex(p) {
+    return !!fs.statSync(p);
+  }
+  function sel() {
+    try {
+      let n = JSON.parse(fs.readFileSync(S)).startup | 0;
+      return n >= -2 && n < 9 ? n : -2;
+    } catch (e) {
+      return -2;
+    }
+  }
+  function saveSel(n) {
+    try {
+      fs.writeFileSync(S, '{"startup":' + n + '}');
+      return 1;
+    } catch (e) {
+      return 0;
+    }
+  }
+  function vf(n) {
+    switch (n) {
+      case 0:
+        return B + 'MISTER.AVI';
+      case 1:
+        return B + 'VAULTGIRL.AVI';
+      case 2:
+        return B + 'DEATHCLAW.AVI';
+      case 3:
+        return B + 'YESMAN.AVI';
+      case 4:
+        return B + 'ENCLAVE.AVI';
+      case 5:
+        return B + 'BOS.AVI';
+      case 6:
+        return B + 'MOTHMAN.AVI';
+      case 7:
+        return B + 'MINUTEMEN.AVI';
+      case 8:
+        return B + 'ENCLAVE_PIPBOY.AVI';
+    }
+  }
+  function af(n) {
+    switch (n) {
+      case 0:
+        return B + 'MISTER.WAV';
+      case 1:
+        return B + 'VAULTGIRL.WAV';
+      case 2:
+        return B + 'DEATHCLAW.WAV';
+      case 3:
+        return B + 'YES MAN.WAV';
+      case 4:
+        return B + 'THE ENCLAVE.WAV';
+      case 5:
+        return B + 'BOS.WAV';
+      case 6:
+        return B + 'MOTHMAN.WAV';
+      case 7:
+        return B + 'MINUTEMEN.WAV';
+      case 8:
+        return B + 'ENCLAVE_PIPBOY.WAV';
+    }
+  }
+  function fm(n) {
+    if (n === 8) return 25500;
+    if (n === 0) return 5930;
+    if (n === 3) return 4597;
+    if (n >= 4 && n <= 7) return 5263;
+    return 3180;
+  }
+  function hm(n) {
+    if (n === 8) return 30000;
+    if (n === 0) return 9445;
+    if (n === 3) return 8112;
+    if (n >= 4 && n <= 7) return 8778;
+    return 6695;
+  }
+  function valid(n) {
+    return n >= 0 && n < 9 && ex(vf(n)) && ex(af(n));
+  }
+  function stockBoot() {
+    let m = process.memory(false);
+    h.clear();
+    return Pip.typeText(
+      '\n\n§§§*************** PIP-OS(R) V5.0.1.4 ***************\n\n' +
+        'COPYRIGHT 2068 ROBCO(R) §\n' +
+        'LOADER V' +
+        VERSION +
+        '\n' +
+        'EXEC VERSION ' +
+        process.env.VERSION +
+        ' §\n' +
+        ((m.total * m.blocksize) / 1e3).toFixed(0) +
+        'K RAM SYSTEM\n' +
+        m.free * m.blocksize +
+        ' BYTES FREE\n' +
+        'NO HOLOTAPE FOUND §\n' +
+        'LOAD ROM(1): DEITRIX 2040... COMPLETE §\n' +
+        '\n\n\n\n\n\n\n\n\n\n\n\n\n',
+      40,
+      0,
+      400,
+      240,
+    );
+  }
+  function originalAudio() {
+    return (
+      Pip.__SA22CleanAudio ||
+      Pip.__startupAnimationsOriginalAudioStartV5 ||
+      Pip.__startupAnimationsOriginalAudioStartV6 ||
+      Pip.__startupAnimationsOriginalAudioStartV7 ||
+      Pip.__startupAnimationsOriginalAudioStartV8 ||
+      Pip.__startupAnimationsStockAudioStartV9 ||
+      Pip.__startupAnimationsStockAudioStartV12 ||
+      Pip.__startupAnimationsStockAudioV14 ||
+      Pip.audioStart.bind(Pip)
+    );
+  }
+  function originalBoot() {
+    return Pip.__SA22CleanBoot || stockBoot;
+  }
+  function clearFlags() {
+    delete global.__startupAnimationsHoloLiveV5;
+    delete global.__startupAnimationsHoloLiveV6;
+    delete global.__startupAnimationsHoloLiveV7;
+    delete global.__startupAnimationsHoloLiveV8;
+    delete global.__startupAnimationsHoloLiveV9;
+    delete global.__startupAnimationsHoloLiveV12;
+    delete global.__startupAnimationsHoloLiveV14;
+    delete global.__SA16;
+    delete global.__SA17;
+    delete global.__SA18;
+    delete global.__SA22;
+    delete global.__SA18CRT;
+    delete global.__SA16C;
+    delete global.__startupAnimationsSuppressCRT12;
+    delete global.__startupAnimationsSuppressCRT14;
+  }
+  function restore() {
+    let q = originalAudio(),
+      b = originalBoot();
+    if (t) {
+      clearTimeout(t);
+      t = undefined;
+    }
+    if (ownVideo) {
+      Pip.videoStop();
+      ownVideo = 0;
+    }
+    if (ownAudio) {
+      Pip.audioStop();
+      ownAudio = 0;
+    }
+    clearFlags();
+    if (aHook && Pip.audioStart === aHook && q) Pip.audioStart = q;
+    if (bHook && Pip.bootAnimation === bHook) Pip.bootAnimation = b;
+    aHook = undefined;
+    bHook = undefined;
+  }
+  function hook() {
+    clearFlags();
+    let q = originalAudio();
+    Pip.__SA18A = q;
+    global.__SA22 = 1;
+    global.__SA18CRT = 0;
+    aHook = function (p) {
+      let n = sel();
+      if (!valid(n)) {
+        restore();
+        return q.apply(Pip, arguments);
+      }
+      if (global.__SA18CRT && p === 'SOUND/FX/CRT_ON2.WAV') {
+        global.__SA18CRT = 0;
+        return;
+      }
+      return q.apply(Pip, arguments);
+    };
+    Pip.audioStart = aHook;
+    bHook = function () {
+      let n = sel();
+      if (!valid(n)) {
+        restore();
+        return stockBoot();
+      }
+      return new Promise(function (done) {
+        let a,
+          b,
+          d,
+          e,
+          x = 0;
+        function end() {
+          if (x) return;
+          x = 1;
+          if (a) clearTimeout(a);
+          if (b) clearTimeout(b);
+          if (d) clearTimeout(d);
+          if (e) clearTimeout(e);
+          if (ownVideo) {
+            Pip.videoStop();
+            ownVideo = 0;
+          }
+          if (ownAudio) {
+            Pip.audioStop();
+            ownAudio = 0;
+          }
+          h.reset();
+          global.__SA18CRT = 1;
+          setTimeout(done, 160);
+        }
+        function play() {
+          if (x) return;
+          if (!valid(n)) {
+            x = 1;
+            restore();
+            try {
+              stockBoot().then(done);
+            } catch (e) {
+              done();
+            }
+            return;
+          }
+          Pip.videoStop();
+          Pip.audioStop();
+          h.clear();
+          h.flip();
+          function startVideo() {
+            if (x) return;
+            E.defrag();
+            try {
+              Pip.videoStart(vf(n), { x: 0, y: 0 });
+              ownVideo = 1;
+            } catch (e) {
+              restore();
+              x = 1;
+              try {
+                stockBoot().then(done);
+              } catch (z) {
+                done();
+              }
+              return;
+            }
+            e = setTimeout(
+              function () {
+                e = undefined;
+                if (x) return;
+                try {
+                  Pip.audioStart(af(n));
+                  ownAudio = 1;
+                } catch (z) {
+                  ownAudio = 0;
+                }
+              },
+              n === 8 ? 420 : 180,
+            );
+            b = setTimeout(end, fm(n) + (n === 8 ? 420 : 180));
+          }
+          if (n === 8) {
+            e = setTimeout(function () {
+              e = undefined;
+              startVideo();
+            }, 260);
+            return;
+          }
+          startVideo();
+        }
+        try {
+          a = setTimeout(play, sel() === 8 ? 2150 : 1895);
+          d = setTimeout(end, hm(n));
+        } catch (e) {
+          restore();
+          try {
+            stockBoot().then(done);
+          } catch (z) {
+            done();
+          }
+        }
+      });
+    };
+    Pip.bootAnimation = bHook;
+  }
+  function header() {
+    h.clear().setColor(3).setFontMonofonto16().setFontAlign(0, -1);
+    h.drawImage(TI.title, (480 - TI.title.width) >> 1, 6);
+  }
+  function scanInstalled() {
+    let n;
+    for (n = 0; n < 9; n++) ic[n] = valid(n) ? 1 : 0;
+  }
+  function installed(n) {
+    return n >= 0 && n < 9 && ic[n];
+  }
+  function specialEntries() {
+    let a = [];
+    if (installed(0)) a.push(['Mister Handy', 0]);
+    if (installed(1)) a.push(['Vault Girl', 1]);
+    if (installed(2)) a.push(['Deathclaw Vault Experiment', 2]);
+    if (installed(8)) a.push(['Enclave PIP-BOY', 8]);
+    return a;
+  }
+  function factionEntries() {
+    let a = [];
+    if (installed(4)) a.push(['The Enclave', 4]);
+    if (installed(5)) a.push(['The Brotherhood of Steel', 5]);
+    if (installed(7)) a.push(['The Minutemen', 7]);
+    if (installed(6)) a.push(['Mothman', 6]);
+    return a;
+  }
+  function npcEntries() {
+    let a = [];
+    if (installed(3)) a.push(['YES MAN', 3]);
+    return a;
+  }
+  function mainEntries() {
+    let a = [['Default Bootup', -1, 0]];
+    if (specialEntries().length) a.push(['Special Bootups', -10, 1]);
+    if (factionEntries().length) a.push(['Faction Bootups', -11, 2]);
+    if (npcEntries().length) a.push(['NPC Bootups', -12, 3]);
+    a.push(['< Back', -99, -1]);
+    return a;
+  }
+  function pageEntries() {
+    if (page === 1) return specialEntries();
+    if (page === 2) return factionEntries();
+    if (page === 3) return npcEntries();
+    return mainEntries();
+  }
+  function draw() {
+    let y = 112,
+      n,
+      a = pageEntries();
+    header();
+    if (page === 0) {
+      for (n = 0; n < a.length; n++) {
+        if (c === n) Pip.shadeBox(24, y - 4, 456, y + 22);
+        h.setColor(3).setFontAlign(-1, -1).drawString(a[n][0], 38, y);
+        if (a[n][1] === -1 && s === -1)
+          h.setFontAlign(1, -1).drawString('ACTIVE', 442, y);
+        y += 36;
+      }
+    } else {
+      for (n = 0; n < a.length; n++) {
+        if (c === n) Pip.shadeBox(24, y - 4, 456, y + 22);
+        h.setColor(3).setFontAlign(-1, -1).drawString(a[n][0], 38, y);
+        if (s === a[n][1]) h.setFontAlign(1, -1).drawString('ACTIVE', 442, y);
+        y += 36;
+      }
+      if (c === a.length) Pip.shadeBox(24, y - 4, 456, y + 22);
+      h.setColor(3).setFontAlign(-1, -1).drawString('< Back', 38, y);
+    }
+  }
+  function maxRow() {
+    let a = pageEntries();
+    return page === 0 ? a.length - 1 : a.length;
+  }
+  function activateStartup(id) {
+    if (!saveSel(id)) return;
+    s = id;
+    if (t) clearTimeout(t);
+    t = setTimeout(function () {
+      t = undefined;
+      hook();
+    }, 0);
+    draw();
+  }
+  function knob(d) {
+    if (d) {
+      c += d > 0 ? 1 : -1;
+      let m = maxRow();
+      if (c < 0) c = m;
+      if (c > m) c = 0;
+      Pip.playSound('SCROLL');
+      draw();
+      return;
+    }
+    Pip.playSound('SELECT');
+    if (page === 0) {
+      let a = mainEntries(),
+        e = a[c];
+      if (e[1] === -99) {
+        Pip.changeMenu('MISC.JS');
+        return;
+      }
+      if (e[1] === -1) {
+        if (!saveSel(-1)) return;
+        s = -1;
+        restore();
+        draw();
+        return;
+      }
+      page = e[2];
+      c = 0;
+      draw();
+      return;
+    }
+    let a = pageEntries();
+    if (c === a.length) {
+      let old = page;
+      page = 0;
+      c = 0;
+      let m = mainEntries(),
+        i;
+      for (i = 0; i < m.length; i++)
+        if (m[i][2] === old) {
+          c = i;
+          break;
+        }
+      draw();
+      return;
+    }
+    if (c < a.length) activateStartup(a[c][1]);
+  }
+  s = sel();
+  scanInstalled();
+  if (s >= 0 && !valid(s)) {
+    saveSel(-1);
+    s = -1;
+  }
+  if (s === -2) s = -1;
+  c = 0;
+  Pip.onExclusive('knob1', knob);
+  draw();
+  if (s < 0) {
+    restore();
+  } else {
+    t = setTimeout(function () {
+      t = undefined;
+      hook();
+    }, 0);
+  }
+  return {
+    id: 'STARTUPANIMATIONS',
+    notDefault: true,
+    fullscreen: true,
+    remove: function () {
+      if (t) clearTimeout(t);
+      Pip.removeListener('knob1', knob);
+      if (ownVideo) {
+        Pip.videoStop();
+        ownVideo = 0;
+      }
+      if (ownAudio) {
+        Pip.audioStop();
+        ownAudio = 0;
+      }
+      h.clear();
+    },
+  };
+});
