--- holotapes/PipCoStartupSystems/assets/TITLE_IMG.JS
@@ -1,11 +1,0 @@
-({
-  title: {
-    width: 217,
-    height: 41,
-    bpp: 1,
-    transparent: 0,
-    buffer: atob(
-      'AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAfgAAAAAAAAAAAAAAAH4AAAAAAAAAAAAAAAAA44AAAAAAAAAAAAAAAOOAAAAAAAAAAAAAAAAA4OAAAAAAAAAAAAAAAODgAAAAAAAAAAAAAAAA4DgAAAAAAAAAAAAAAOA4AAAAAAAAAAAAAAAAcBwAAAAAAAAAAAAAAHAcAAAAAAAAAAAAAAAAOA4cAAAAAHAAAAAAADgOAAAAAcAAAAAAAAAAPAcOAAAAADgAAAAAADwHAAAAAOAAAAAAAAAAHgcHAAAAABwAAAAAAB4HAAAAAHAAAAAAAAAADwADgAAAAA4AAAAAAA8AAAAAADgAAAAAAAAAA8ABwAAAAAcAAAOAAAPAAAAAHBwAAAAAAAOAAfgD/A94OO/zg8HPAAH4AcHgHj/AfDng8APAAP8AcB/8HfnBweD/4AD/AODwDwcA/x39/AHgAD/gOB8eDvzg4PB88AA/4HB4B8OA8c/+/wD4AAf4HA8PB8xwcHg8PAAH+Dg8B+HAcGfHx4D8AAD+Dg8Hg8A4ODwcDgAA/hweA/DgeDPDwcB+AAAfhweDwcAcHB4OB4AAH44PAThwPDnh4OAnAAAH44OB4OAODg8HA8AAB+cHgZ44HBzg8HgzwAPB8cPA8HAHBweDgeADwfODwM8cHhxweDwZ4APg+OHgeDgDg4PBwPAD4PnB4EeODxw4PBwI8AHgPHDwPBwBwcHg4HgB4Dzg8GHHB/wcHg4MOAHgHjh4Hg4A4ODwcDwB4B5weDDzg/AODwcGHgDwDxw8Dw8AcHB4eB4A8A84PDx5weAHB4OHjwB4B44eB4eAODg8PA8AeAecHhw94PAHg8HDh4A8B8cPg8fAHDwePg8APAfODwwf8PwHweDhg8APA8PHw+PgDx8fHweADwPHj44fvF8G4PB5w+AHw8H+//9wB//7++eAB8PD/8f/H+f+cHg//+AB/8B+f5+4Afn8+f+AAf/A/+z/B+H+ODwP3+AAH4AeDwccAHh8OOeAAB+APvwfAeB8HB4Dw+AAAAAAAAAAAAAAAHAAAAAAAHgAAAAAAAAAAAAAAAAAAAAAAAAAADgAAAAAAPwAAAAAAAAAAAAAAAAAAAAAAAAAABwAAAAAAN4AAAAAAAAAAAAAAAAAAAAAAAAAAA4AAAAAAM8AAAAAAAAAAAAAAAAAAAAAAAAAAAcAAAAAAEeAAAAAAAAAAAAAAAAAAAAAAAAAAAOAAAAAAGOAAAAAAAAAAAAAAAAAAAAAAAAAAAHAAAAAABHAAAAAAAAAAAAAAAAAAAAAAAAAAADgAAAAAA/gAAAAAAAAAAAAAAAAAAAAAAAAAABgAAAAAAPAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA',
-    ),
-  },
-});
