+++ holotapes/PipCoStartupSystems/metadata.json
@@ -2,16 +2,16 @@
   "id": "startup-animations",
   "name": "PIP-CO Startup Systems",
   "author": "@LlamaYeYe",
-  "version": "1.2.0",
+  "version": "1.1.0",
   "description": "Choose custom startup animations and audio for the Wand Company Pip-Boy 3000. Install only the startup packs you want; the holotape automatically shows only the media currently installed.",
-  "icon": "assets/PIP_CO_STARTUP_SYSTEMS.IMG",
+  "icon": "assets/PIP_CO_STARTUP_SYSTEMS.png",
   "previews": ["previews/menu.bmp"],
   "type": "app",
   "readme": "README.md",
   "storage": [
     {
       "name": "HOLO/STARTUP_ANIMATIONS/APP.JS",
-      "url": "assets/APP.MIN.JS"
+      "url": "app.min.js"
     },
     {
       "name": "HOLO/STARTUP_ANIMATIONS/SELECT.JSON",
@@ -20,108 +20,116 @@
     {
       "name": "HOLO/STARTUP_ANIMATIONS/TITLE.BIN",
       "url": "assets/TITLE.BIN"
-    },
-    {
-      "name": "HOLO/STARTUP_ANIMATIONS/TITLE_IMG.JS",
-      "url": "assets/TITLE_IMG.JS"
     }
   ],
   "storageOptional": [
     {
-      "label": "Mister Handy \u2014 Animation",
+      "label": "Mister Handy — Animation",
       "name": "HOLO/STARTUP_ANIMATIONS/MISTER.AVI",
       "sizeKB": 942,
       "url": "assets/MISTER.AVI"
     },
     {
-      "label": "Mister Handy \u2014 Audio",
+      "label": "Mister Handy — Audio",
       "name": "HOLO/STARTUP_ANIMATIONS/MISTER.WAV",
       "sizeKB": 180,
       "url": "assets/MISTER.WAV"
     },
     {
-      "label": "Vault Girl \u2014 Animation",
+      "label": "Vault Girl — Animation",
       "name": "HOLO/STARTUP_ANIMATIONS/VAULTGIRL.AVI",
-      "sizeKB": 454,
+      "sizeKB": 453,
       "url": "assets/VAULTGIRL.AVI"
     },
     {
-      "label": "Vault Girl \u2014 Audio",
+      "label": "Vault Girl — Audio",
       "name": "HOLO/STARTUP_ANIMATIONS/VAULTGIRL.WAV",
       "sizeKB": 92,
       "url": "assets/VAULTGIRL.WAV"
     },
     {
-      "label": "Deathclaw Vault Experiment \u2014 Animation",
+      "label": "Deathclaw Vault Experiment — Animation",
       "name": "HOLO/STARTUP_ANIMATIONS/DEATHCLAW.AVI",
-      "sizeKB": 746,
+      "sizeKB": 745,
       "url": "assets/DEATHCLAW.AVI"
     },
     {
-      "label": "Deathclaw Vault Experiment \u2014 Audio",
+      "label": "Deathclaw Vault Experiment — Audio",
       "name": "HOLO/STARTUP_ANIMATIONS/DEATHCLAW.WAV",
       "sizeKB": 92,
       "url": "assets/DEATHCLAW.WAV"
     },
     {
-      "label": "YES MAN \u2014 Animation",
+      "label": "YES MAN — Animation",
       "name": "HOLO/STARTUP_ANIMATIONS/YESMAN.AVI",
-      "sizeKB": 52,
+      "sizeKB": 51,
       "url": "assets/YESMAN.AVI"
     },
     {
-      "label": "YES MAN \u2014 Audio",
+      "label": "YES MAN — Audio",
       "name": "HOLO/STARTUP_ANIMATIONS/YES MAN.WAV",
-      "sizeKB": 138,
+      "sizeKB": 137,
       "url": "assets/YES MAN.WAV"
     },
     {
-      "label": "The Enclave \u2014 Animation",
+      "label": "The Enclave — Animation",
       "name": "HOLO/STARTUP_ANIMATIONS/ENCLAVE.AVI",
-      "sizeKB": 114,
+      "sizeKB": 113,
       "url": "assets/ENCLAVE.AVI"
     },
     {
-      "label": "The Enclave \u2014 Audio",
+      "label": "The Enclave — Audio",
       "name": "HOLO/STARTUP_ANIMATIONS/THE ENCLAVE.WAV",
-      "sizeKB": 159,
+      "sizeKB": 158,
       "url": "assets/THE ENCLAVE.WAV"
     },
     {
-      "label": "The Brotherhood of Steel \u2014 Animation",
+      "label": "The Brotherhood of Steel — Animation",
       "name": "HOLO/STARTUP_ANIMATIONS/BOS.AVI",
-      "sizeKB": 870,
+      "sizeKB": 869,
       "url": "assets/BOS.AVI"
     },
     {
-      "label": "The Brotherhood of Steel \u2014 Audio",
+      "label": "The Brotherhood of Steel — Audio",
       "name": "HOLO/STARTUP_ANIMATIONS/BOS.WAV",
       "sizeKB": 157,
       "url": "assets/BOS.WAV"
     },
     {
-      "label": "Mothman \u2014 Animation",
+      "label": "Mothman — Animation",
       "name": "HOLO/STARTUP_ANIMATIONS/MOTHMAN.AVI",
       "sizeKB": 774,
       "url": "assets/MOTHMAN.AVI"
     },
     {
-      "label": "Mothman \u2014 Audio",
+      "label": "Mothman — Audio",
       "name": "HOLO/STARTUP_ANIMATIONS/MOTHMAN.WAV",
-      "sizeKB": 159,
+      "sizeKB": 158,
       "url": "assets/MOTHMAN.WAV"
     },
     {
-      "label": "The Minutemen \u2014 Animation",
+      "label": "The Minutemen — Animation",
       "name": "HOLO/STARTUP_ANIMATIONS/MINUTEMEN.AVI",
-      "sizeKB": 35,
+      "sizeKB": 34,
       "url": "assets/MINUTEMEN.AVI"
     },
     {
-      "label": "The Minutemen \u2014 Audio",
+      "label": "The Minutemen — Audio",
       "name": "HOLO/STARTUP_ANIMATIONS/MINUTEMEN.WAV",
-      "sizeKB": 159,
+      "sizeKB": 158,
       "url": "assets/MINUTEMEN.WAV"
+    },
+    {
+      "label": "Enclave PIP-BOY — Animation",
+      "name": "HOLO/STARTUP_ANIMATIONS/ENCLAVE_PIPBOY.AVI",
+      "sizeKB": 5350,
+      "url": "assets/ENCLAVE_PIPBOY.AVI"
+    },
+    {
+      "label": "Enclave PIP-BOY — Audio",
+      "name": "HOLO/STARTUP_ANIMATIONS/ENCLAVE_PIPBOY.WAV",
+      "sizeKB": 797,
+      "url": "assets/ENCLAVE_PIPBOY.WAV"
     }
   ]
 }
